Refactor: move async-runtime to dir `type_config/async_runtime/`
