+++ openraft/src/impls/mod.rs
@@ -1,7 +1,7 @@
 //! Collection of implementations of usually used traits defined by Openraft
 
-pub use crate::async_runtime::TokioRuntime;
 pub use crate::entry::Entry;
 pub use crate::node::BasicNode;
 pub use crate::node::EmptyNode;
 pub use crate::raft::responder::impls::OneshotResponder;
+pub use crate::type_config::async_runtime::impls::TokioRuntime;
